requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
psycopg2==2.9.4
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
django-filter==22.1
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4 # Libreary for swagger and redoc
